requirements.txt
openai==0.28.1
faiss-cpu
numpy
pandas
python-dotenv
streamlit
tqdm
PyMuPDF
python-docx
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
PyPDF2
openpyxl
